projects/genomics/variant-analysis/tier-2/requirements.txt
# AWS SDK and utilities
boto3>=1.26.0
botocore>=1.29.0
awscli>=1.27.0

# Data analysis and visualization
pandas>=1.4.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0

# Genomics
pysam>=0.20.0
vcf>=0.6.10
Biopython>=1.80

# Jupyter and development
jupyter>=1.0.0
jupyterlab>=3.5.0
ipywidgets>=8.0.0

# Testing and quality
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Utilities
python-dotenv>=0.20.0
requests>=2.28.0
tqdm>=4.64.0
